file prep for gemoma

+++ GeMoMa/exon_info_to_gff.py
@@ -1,13 +1,10 @@
+import os.path
 import sys
 
 from Basic_Tools.lists_and_files import file_to_list
 
-if __name__ == "__main__":
-
-    exon_file = "/Users/tonyx/Documents/chang_lab/good_shark_references/GUCY2F/Carcharodon carcharias/Carcharodon carcharias/3249_XM_041199335.1.fas"
-    gff_file = "/Users/tonyx/Documents/chang_lab/gffs/gucy2f.gff"
-    fasta_file = "/Users/tonyx/Documents/chang_lab/gffs/gucy2f.fas"
-
+def exon_file_to_gff(exon_file: str, gff_file: str, fasta_file: str):
+    # if we don't have a single fasta entry, leave (that's not going to be the case though, is it?
     exon_file_array = file_to_list(exon_file)
     if len(exon_file_array) < 2:
         print("not enough lines in the exon file")
@@ -17,6 +14,8 @@ if __name__ == "__main__":
     if header[0] != ">":
         print("not in fasta heading")
         sys.exit()
+
+    # TODO add back if mRNA is needed
 
     mrna_pos = -1
 
@@ -29,17 +28,22 @@ if __name__ == "__main__":
         print("no mrna name")
         sys.exit()
 
+    # the only reason I'm getting the mRNA position is bc it's relative tot he CDS
     cds_pos = mrna_pos + 2
 
     mrna_name = header.split()[mrna_pos][5:]
     gene_name = header.split()[0][1:]
 
-    gff_f = open(gff_file, "w")
-    gff_f.write("##gff-version 3\n")
+    if not os.path.isfile(gff_file):
+        gff_f = open(gff_file, "w")
+        gff_f.write("##gff-version 3\n")
+    else:
+        gff_f = open(gff_file, "a")
 
+    # just get the cds_ranges and sequence from the file
     cds_ranges = []
-
     seq = ""
+
     for i in range(len(exon_file_array)):
         if exon_file_array[i][0] == ">":
             cds = exon_file_array[i].split()[cds_pos].split("-")
@@ -47,17 +51,10 @@ if __name__ == "__main__":
         else:
             seq += exon_file_array[i].strip()
 
+    # the length of the coding sequence is the last one
     cds_len = cds_ranges[-1][1]
-    gff_f.write(gene_name + "\t" +
-                "." + "\t" +
-                "gene" + "\t" +
-                "1" + "\t" +
-                cds_len + "\t" +
-                "." + "\t" +
-                "+" + "\t" +
-                "." + "\t" +
-                "ID=gene-" + gene_name +"\n")
 
+    # technically in the mRNA category (seems like this is all we need though, LOL)
     gff_f.write(gene_name + "\t" +
                 "." + "\t" +
                 "mRNA" + "\t" +
@@ -66,7 +63,7 @@ if __name__ == "__main__":
                 "." + "\t" +
                 "+" + "\t" +
                 "." + "\t" +
-                "ID=rna-" + mrna_name + "\n")
+                "ID=rna-" + gene_name + "\n")
 
     for cds in cds_ranges:
         gff_f.write(gene_name + "\t" +
@@ -79,12 +76,18 @@ if __name__ == "__main__":
                     "." + "\t" +
                     "Parent=rna-" + mrna_name + "\n")
 
-    gff_f.write("###")
-
-    fasta_f = open(fasta_file, "w")
+    fasta_f = open(fasta_file, "a")
     fasta_f.write(">" + gene_name + " " + mrna_name + "\n")
-    fasta_f.write(seq)
+    fasta_f.write(seq + "\n")
 
+
+if __name__ == "__main__":
+
+    exon_file = "/Users/tonyx/Documents/chang_lab/converted_NEPR5/RARA/Carcharodon carcharias.fas"
+    gff_file = "/Users/tonyx/Documents/chang_lab/gffs/RARA_test.gff"
+    fasta_file = "/Users/tonyx/Documents/chang_lab/gffs/RARA_test.fas"
+
+    exon_file_to_gff(exon_file,gff_file,fasta_file)
